move / util 파일 옮기기 및 타입 이름 수정 및 home 디렉토리 url에서 나타나지 않게 수정 #3
