+++ src/app/(home)/types/HomePost.ts
