+++ src/utils/home-emotion.util.ts
